warnings--
svn path=/trunk/extragear/office/tellico/; revision=1079608

+++ src/fetch/z3950connection.cpp
@@ -467,7 +467,7 @@ QByteArray Z3950Connection::iconvRun(const QByteArray& text_, const QString& fro
   yaz_iconv(cd, 0, 0, &result, &outlen);
 
   // length is pointer difference
-  int len = result - result0.data();
+  ptrdiff_t len = result - result0.data();
 
   QByteArray output(result0.data(), len+1);
 //  myDebug() << "-------------------------------------------";
